remove post on type assertions and non-null assertions
The feedback on this post was overwhelmingly negative. The post advocated for the use
of non-null assertions over type assertions, which none of the feedback strictly
disagreed with. However, the feedback was that assertions should generally be avoided
in favor of run-time checks, and that the post advocating for the use of non-null
assertions is bad advice.

After some reflection, I agree that the post is not helpful. It's correct in that
non-null assertions are preferable to type assertions, but the post should have
discouraged readers from using assertions generally. As one commenter pointed out, the
post is technically correct but misses the point.

I'm opting to delete the post. I could have decided to rework the post to
discourage assertions generally, but I think the post is structured around a
flawed premise and should be scrapped entirely.